Removing the closed field from the ElasticsearchTransportClient in favour of the client's connectedNodes() list.  This list is managed internally by the client's node service when sniff=true.
To ensure that a node that drops and comes back is picked up by log4j2,
I removed the sniff plugin attribute and set it to true, always.
Without detecting when a node comes back, a client that fails to write
when the node goes down will never send log events upstream.